vote-5

+++ src/main/java/dev/marioszocs/hotelreservationapi/serviceImp/ReservationServiceImp.java
@@ -3,6 +3,7 @@ package dev.marioszocs.hotelreservationapi.serviceImp;
 import dev.marioszocs.hotelreservationapi.constants.ErrorMessages;
 import dev.marioszocs.hotelreservationapi.dto.IdEntity;
 import dev.marioszocs.hotelreservationapi.dto.SuccessEntity;
+import dev.marioszocs.hotelreservationapi.entity.Hotel;
 import dev.marioszocs.hotelreservationapi.entity.Reservation;
 import dev.marioszocs.hotelreservationapi.exception.InvalidRequestException;
 import dev.marioszocs.hotelreservationapi.repository.HotelRepository;
@@ -11,6 +12,7 @@ import dev.marioszocs.hotelreservationapi.service.ReservationService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import jakarta.transaction.Transactional;
 import java.text.ParseException;
@@ -44,8 +46,8 @@ public class ReservationServiceImp implements ReservationService {
      */
     @Override
     public Reservation getReservation(Integer id) {
-        validateReservationExistence(id);
-        return reservationRepository.findById(id).get();
+        return reservationRepository.findById(id)
+                .orElseThrow(() -> new InvalidRequestException(ErrorMessages.INVALID_ID_EXISTENCE));
     }
 
     /**
@@ -55,27 +57,27 @@ public class ReservationServiceImp implements ReservationService {
      */
     @Override
     public IdEntity saveReservation(Reservation reservations) {
-        Integer reservationsInventoryId = reservations.getHotelId();
+        Integer hotelId = reservations.getHotelId();
+        
+        Hotel hotel = hotelRepository.findById(hotelId)
+                .orElseThrow(() -> new InvalidRequestException(ErrorMessages.INVALID_HOTEL_IN_RESERVATION));
 
-        //boolean to determine if the Reservation is valid through the existence of the inventory ID.
-        //if the inventory ID exists, then continue
-        if (validateHotelExistenceById(reservationsInventoryId)) {
-            //boolean to determine if there is already a pre-existing reservation that overlaps with the users
-            if (reservationOverlaps(reservations)) {
-                throw new InvalidRequestException(ErrorMessages.INVALID_DATE_OVERLAP);
-            }
-            //determine if the dates are out of the Inventory's bounds  //TODO: getById is deprecated
-            if (dateIsBefore(hotelRepository.getById(reservations.getHotelId()).getAvailableFrom(), reservations.getCheckIn()) && dateIsBefore(reservations.getCheckOut(), hotelRepository.getById(reservations.getHotelId()).getAvailableTo())) {
-                reservations = reservationRepository.save(reservations);
-                IdEntity idEntity = new IdEntity();
-                idEntity.setId(reservations.getId());
-                return idEntity;
-            } else {
-                throw new InvalidRequestException(ErrorMessages.INVALID_RESERVATION_DATES);
-            }
+        if (!StringUtils.hasText(hotel.getAvailableFrom()) || !StringUtils.hasText(hotel.getAvailableTo())) {
+            throw new InvalidRequestException(ErrorMessages.EMPTY_HOTEL_DATES);
+        }
+
+        if (reservationRepository.existsOverlappingReservation(hotelId, reservations.getCheckIn(), reservations.getCheckOut())) {
+            throw new InvalidRequestException(ErrorMessages.INVALID_DATE_OVERLAP);
+        }
+
+        if (reservations.getCheckIn().compareTo(hotel.getAvailableFrom()) >= 0 
+                && reservations.getCheckOut().compareTo(hotel.getAvailableTo()) <= 0) {
+            Reservation savedReservation = reservationRepository.save(reservations);
+            IdEntity idEntity = new IdEntity();
+            idEntity.setId(savedReservation.getId());
+            return idEntity;
         } else {
-            //Throw error if the Inventory ID does not exist
-            throw new InvalidRequestException(ErrorMessages.INVALID_HOTEL_IN_RESERVATION);
+            throw new InvalidRequestException(ErrorMessages.INVALID_RESERVATION_DATES);
         }
     }
 
@@ -87,10 +89,12 @@ public class ReservationServiceImp implements ReservationService {
      */
     @Override
     public SuccessEntity deleteReservation(Integer id) {
-        validateReservationExistence(id);
+        if (!reservationRepository.existsById(id)) {
+            throw new InvalidRequestException(ErrorMessages.INVALID_ID_EXISTENCE);
+        }
         reservationRepository.deleteById(id);
         SuccessEntity successEntity = new SuccessEntity();
-        successEntity.setSuccess(!reservationRepository.existsById(id));
+        successEntity.setSuccess(true);
         return successEntity;
     }
 
@@ -101,14 +105,13 @@ public class ReservationServiceImp implements ReservationService {
      */
     @Override
     public boolean validateHotelExistenceById(Integer id) {
-        if (!hotelRepository.existsById(id)) {
-            throw new InvalidRequestException(ErrorMessages.INVALID_ID_EXISTENCE);
-        } else if (hotelRepository.getById(id).getAvailableFrom() == null && hotelRepository.getById(id).getAvailableTo() == null) {  //TODO getById is deprecated
-            //Checks if the inventory has available to and available from dates, if not then throw an error as a reservation cannot be made.
+        Hotel hotel = hotelRepository.findById(id)
+                .orElseThrow(() -> new InvalidRequestException(ErrorMessages.INVALID_ID_EXISTENCE));
+        
+        if (!StringUtils.hasText(hotel.getAvailableFrom()) || !StringUtils.hasText(hotel.getAvailableTo())) {
             throw new InvalidRequestException(ErrorMessages.EMPTY_HOTEL_DATES);
-        } else {
-            return true;
         }
+        return true;
     }
 
     /**
@@ -120,12 +123,7 @@ public class ReservationServiceImp implements ReservationService {
      */
     @Override
     public boolean dateIsBefore(String date1, String date2) {
-        SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd");
-        try {
-            return simpleDateFormat.parse(date1).before(simpleDateFormat.parse(date2));
-        } catch (ParseException e) {
-            throw new InvalidRequestException(ErrorMessages.PARSE_ERROR);
-        }
+        return date1.compareTo(date2) < 0;
     }
 
     /**
@@ -136,29 +134,10 @@ public class ReservationServiceImp implements ReservationService {
      */
     @Override
     public boolean reservationOverlaps(Reservation reservations) {
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
-
-        return reservationRepository.findAll().stream().anyMatch(dataBaseRes -> {
-            //If two reservations have the same inventory id, then compare their check in and checkout dates
-            if (dataBaseRes.getHotelId() == reservations.getHotelId()) {
-                try {
-                    int checkInBeforeDbCheckOut = sdf.parse(reservations.getCheckIn()).compareTo(sdf.parse(dataBaseRes.getCheckOut()));
-                    int checkOutBeforeDbCheckIn = sdf.parse(reservations.getCheckOut()).compareTo(sdf.parse(dataBaseRes.getCheckIn()));
-                    log.debug("check in int " + checkInBeforeDbCheckOut);
-                    log.debug("check out int " + checkOutBeforeDbCheckIn);
-                    if (checkInBeforeDbCheckOut == 0 || checkOutBeforeDbCheckIn == 0) {
-                        return true;
-                    } else {
-                        return checkInBeforeDbCheckOut != checkOutBeforeDbCheckIn;
-                    }
-                } catch (ParseException e) {
-                    throw new InvalidRequestException(ErrorMessages.PARSE_ERROR);
-                }
-            } else {
-                return false;
-            }
-
-        });
+        return reservationRepository.existsOverlappingReservation(
+                reservations.getHotelId(), 
+                reservations.getCheckIn(), 
+                reservations.getCheckOut());
     }
 
     /**
@@ -169,10 +148,9 @@ public class ReservationServiceImp implements ReservationService {
      */
     @Override
     public boolean validateReservationExistence(Integer id) {
-        if(!reservationRepository.existsById(id)){
+        if (!reservationRepository.existsById(id)) {
             throw new InvalidRequestException(ErrorMessages.INVALID_ID_EXISTENCE);
-        } else {
-            return true;
         }
+        return true;
     }
 }
